ENH "lazy add" for combat and over time buffs. add_xxx() is now very fast, but increment_xxx() isn't short cut as often. for high frequency effects, this is a lot less work, for low frequency it's still beneficial ENH buff manager and hit effects are now "reset" between iterations, instead of "recreating" them, essentially ENH proc data book-keeping is now done in Hit_effect itself

+++ simulator/include/damage_sources.hpp
@@ -46,7 +46,7 @@ struct Damage_sources
                execute_damage + deep_wounds_damage + item_hit_effects_damage + sweeping_strikes_damage;
     }
 
-    [[nodiscard]] long sum_counts() const
+    [[nodiscard]] int sum_counts() const
     {
         return white_mh_count + white_oh_count + bloodthirst_count + mortal_strike_count + slam_count +
                overpower_count + heroic_strike_count + cleave_count + whirlwind_count + hamstring_count +
@@ -70,20 +70,20 @@ struct Damage_sources
     double deep_wounds_damage{};
     double item_hit_effects_damage{};
 
-    long white_mh_count{};
-    long white_oh_count{};
-    long slam_count{};
-    long overpower_count{};
-    long bloodthirst_count{};
-    long mortal_strike_count{};
-    long sweeping_strikes_count{};
-    long execute_count{};
-    long heroic_strike_count{};
-    long cleave_count{};
-    long whirlwind_count{};
-    long hamstring_count{};
-    long deep_wounds_count{};
-    long item_hit_effects_count{};
+    int white_mh_count{};
+    int white_oh_count{};
+    int slam_count{};
+    int overpower_count{};
+    int bloodthirst_count{};
+    int mortal_strike_count{};
+    int sweeping_strikes_count{};
+    int execute_count{};
+    int heroic_strike_count{};
+    int cleave_count{};
+    int whirlwind_count{};
+    int hamstring_count{};
+    int deep_wounds_count{};
+    int item_hit_effects_count{};
 
     std::vector<Damage_instance> damage_instances;
 };
